+++ tool/host.dart
@@ -13,16 +13,6 @@ import 'package:stack_trace/stack_trace.dart';
 import 'package:stream_channel/stream_channel.dart';
 import 'package:js/js.dart';
 
-/// A class defined in content shell, used to control its behavior.
-@JS()
-class _TestRunner {
-  external void waitUntilDone();
-}
-
-/// Returns the current content shell runner, or `null` if none exists.
-@JS()
-external _TestRunner get testRunner;
-
 /// A class that exposes the test API to JS.
 ///
 /// These are exposed so that tools like IDEs can interact with them via remote
@@ -106,10 +96,6 @@ final _currentUrl = Uri.parse(window.location.href);
 /// does mean that the server needs to be sure to nest its [MultiChannel]s at
 /// the same place the client does.
 void main() {
-  // This tells content_shell not to close immediately after the page has
-  // rendered.
-  testRunner?.waitUntilDone();
-
   if (_currentUrl.queryParameters['debug'] == 'true') {
     document.body.classes.add('debug');
   }